service-account/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
loguru
python-dotenv